chatbot/requirements.txt
supabase>=2.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
scikit-learn>=1.2.0
numpy>=1.24.0
xgboost>=2.0.0
fastapi
uvicorn
openai
pydantic
requests
